requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Jinja2==3.1.2
PyJWT==2.8.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
WTForms==3.0.1
alembic==1.12.0
email-validator==2.0.0.post2
pandas==2.1.0
pillow==10.0.0
python-dotenv==1.0.0